% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PakPC2017GUI.R
\name{PakPC2017GUI}
\alias{PakPC2017GUI}
\title{GUI for Pakistan Population Census 2017}
\usage{
PakPC2017GUI(dataset = NA)
}
\arguments{
\item{\code{dataset}}{A dataset (optional).}
}
\value{
A GUI for visualizing data from \code{dataset}.
}
\description{
GUI for Pakistan Population Census 2017.
}
